app/requirements.txt
fastapi
uvicorn
pydantic-settings
pymongo
qdrant-client
sentence-transformers
groq
requests
beautifulsoup4
selenium # For web scraping with Selenium
webdriver-manager # To manage browser drivers for Selenium (useful in Docker too)
langchain-community
langchain-core
python-dotenv
python-multipart # For handling file uploads in FastAPI

